(registration-react) move form-template to components and refactor to stateless
(cherry picked from commit a95e3e4)
