--- applications/registration-react/src/pages/dataset-registration-page/form-template/form-template.scss
@@ -1,3 +1,0 @@
-.fdk-syncError-icon {
-  flex-grow: 1;
-}
